Add "accelerator" strossle widget
