+++ django/apps/core/templates/_strossle.sdk.html
@@ -1,12 +1,5 @@
-<!-- Strossle SDK -->
-<script>
-  window.strossle=window.strossle||function(){(strossle.q=strossle.q||[]).push(arguments)};
-  //strossle('dff15dfe-e8ca-4e6d-b547-8038ab88562b', '.strossle-widget');
-  //strossle('cf6209fc-8eae-4041-bbf9-6d9129336326', '.strossle-widget');
-</script>
+<!-- V1 Strossle SDK-->
+<script async data-spklw-automatic-initialization src="//widgets.sprinklecontent.com/v2/sprinkle.js"></script>
+<!-- V2 Strossle SDK-->
+<script>window.strossle=window.strossle||function(){(strossle.q=strossle.q||[]).push(arguments)}</script>
 <script async src="https://assets.strossle.com/strossle-widget-sdk/1/strossle-widget-sdk.js"></script>
-
-<!-- <script src="//widgets.sprinklecontent.com/v2/sprinkle.js <http://widgets.sprinklecontent.com/v2/sprinkle.js> " async></script> -->
-
-<!-- <script data-spklw-automatic-initialization src="//widgets.sprinklecontent.com/v2/sprinkle.js" async></script>
-End Strossle SDK -->
