+++ webpack/src/universitas/components/RelatedStory/index.js
@@ -1,3 +1,5 @@
 import RelatedStory from './RelatedStory.js'
+import RelatedStories from './RelatedStories.js'
 import './RelatedStory.scss'
 export default RelatedStory
+export { RelatedStories }
